# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "description": "Creates a new user account in the system. The registration data including name, email, password, and other required fields are provided in the request body. The first user to register will automatically be assigned the ADMIN role. Subsequent users will be assigned the EMPLOYEE role by default if no role is specified.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Register a new user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Authenticates a user with their email and password. If successful, returns an access token (JWT) and a refresh token. The access token is used for authenticating subsequent API requests and contains encoded user information. The refresh token can be used to obtain a new access token when the current one expires.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Login with email and password",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/reset-password": {
      "post": {
        "description": "Completes the password reset process by setting a new password for the user. The request body must include the reset token that was sent to the user's email (from the forgot-password endpoint), the user's email address, and the new password. If the token is valid and has not expired, the user's password is updated to the new value.",
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reset password with token",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/2fa/generate": {
      "get": {
        "description": "Generates a new 2FA secret for the authenticated user and returns the secret along with a QR code that can be scanned with authenticator apps. This is the first step in enabling 2FA for a user account. The secret is stored but not activated until the user verifies it with a valid code using the enable-2fa endpoint.",
        "operationId": "AuthController_generate2fa",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate 2FA secret and QR code",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/2fa/enable": {
      "post": {
        "description": "Enables 2FA for the authenticated user after verifying the provided authentication code. The user must have previously generated a 2FA secret using the generate-2fa endpoint. The code provided in the request body must be valid for the stored secret. If successful, 2FA will be enabled for the user account, requiring an additional verification step during future login attempts.",
        "operationId": "AuthController_enable2fa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Enable2faDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Enable 2FA for user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/2fa/disable": {
      "post": {
        "description": "Disables 2FA for the authenticated user. This removes the 2FA requirement for future login attempts. The user will only need their email and password to authenticate after this endpoint is called successfully.",
        "operationId": "AuthController_disable2fa",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disable 2FA for user",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/2fa/verify": {
      "post": {
        "description": "Verifies the 2FA code provided during login for a user with 2FA enabled. This endpoint is called after a successful password authentication when the user has 2FA enabled. The request body must include the temporary token received from the login endpoint and a valid authentication code from the user's authenticator app. If successful, returns the final access and refresh tokens that can be used for API access.",
        "operationId": "AuthController_verify2fa",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify2faDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify 2FA code during login",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "HR Management API",
    "description": "API for managing employees and documents",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Tapiwanashe"
          },
          "lastName": {
            "type": "string",
            "example": "Shoshore"
          },
          "email": {
            "type": "string",
            "example": "tapiwanasheshoshore@gmail.com"
          },
          "employeeNumber": {
            "type": "string",
            "example": "EMP123456"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+263771792254"
          },
          "password": {
            "type": "string",
            "example": "StrongP@ss123",
            "description": "Password must be at least 8 characters long and include at least one uppercase letter, one lowercase letter, one number, and one special character"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "HR_MANAGER",
              "HR_OFFICER",
              "MANAGER",
              "EMPLOYEE"
            ],
            "example": "EMPLOYEE",
            "description": "Role will default to EMPLOYEE if not specified. First user will automatically be ADMIN."
          },
          "contractStartDate": {
            "type": "string",
            "example": "2025-04-27"
          },
          "contractEndDate": {
            "type": "string",
            "example": "2030-04-27"
          },
          "contractType": {
            "type": "string",
            "enum": [
              "Permanent",
              "Temporary",
              "Contract"
            ],
            "example": "Permanent"
          },
          "location": {
            "type": "string",
            "example": "Harare, Zimbabwe"
          },
          "position": {
            "type": "string",
            "example": "Software Engineer"
          },
          "departmentId": {
            "type": "string",
            "example": "550e8400-e29b-41d4-a716-446655440000",
            "description": "UUID of the department. Must be a valid UUID."
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "employeeNumber",
          "phoneNumber",
          "password",
          "contractStartDate"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "Enable2faDto": {
        "type": "object",
        "properties": {
          "twoFactorCode": {
            "type": "string",
            "description": "The verification code from the authenticator app",
            "example": "123456"
          }
        },
        "required": [
          "twoFactorCode"
        ]
      },
      "Verify2faDto": {
        "type": "object",
        "properties": {
          "twoFactorCode": {
            "type": "string",
            "description": "The verification code from the authenticator app",
            "example": "123456"
          },
          "tempToken": {
            "type": "string",
            "description": "The temporary token received after successful password authentication",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "twoFactorCode",
          "tempToken"
        ]
      },
...
    }
  }
}